requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
redis==4.3.4  # https://github.com/redis/redis-py
# hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.2.15  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis

django-htmlmin==0.11.0 # https://django-tailwind.readthedocs.io/en/latest/installation.html

python-webpack-boilerplate==0.0.10
django-admin-honeypot==1.1.0 # https://pypi.org/project/django-admin-honeypot/
django-jet2==1.0.12
# FeedParser==5.2.1
feedparser==6.0.8
google-api-python-client==1.4.1
# integrate htmx with django requests
django-htmx==1.9.0
django-widget-tweaks==1.4.12
django-htmlmin==0.11.0 # https://django-tailwind.readthedocs.io/en/latest/installation.html
crispy-tailwind==0.5.0
